refactor(FNP-71) changed folder structure
