+++ packages/server/utils/index.ts
@@ -1,2 +1,2 @@
-export * from './db';
-export const isDev = process.env.NODE_ENV === 'development';
+export * from './init-postgres';
+export * from './is-dev';
